# contract/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "StitchFlow API",
    "version": "1.1.0",
    "description": "The contract between the StitchFlow apps, the web admin panel and the API. Every operation names the permission action it needs (x-action); a destructive action is super admin only. Dates without a time are \"YYYY-MM-DD\"; instants are ISO-8601 UTC; money is rounded to the paisa. See HANDOVER.md."
  },
  "servers": [
    {
      "url": "https://api.example.com",
      "description": "Replace per environment."
    }
  ],
  "components": {
    "securitySchemes": {
      "bearer": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "UserRole": {
        "type": "string",
        "enum": [
          "superAdmin",
          "admin",
          "viewAdmin",
          "supervisor",
          "worker"
        ],
        "description": "The five fixed roles. `worker` is shown as Operator."
      },
...
      "ShiftScheme": {
        "type": "string",
        "enum": [
          "dayNight",
          "threeShift",
          "manual"
        ]
      },
...
      "ClientKind": {
        "type": "string",
        "enum": [
          "managerApp",
          "crewApp",
          "webAdmin"
        ]
      },
      "AccessLevel": {
        "type": "string",
        "enum": [
          "none",
          "view",
          "edit"
        ],
        "description": "none < view < edit."
      },
      "AccountStatus": {
        "type": "string",
        "enum": [
          "active",
          "disabled"
        ]
      },
      "PropertyStatus": {
        "type": "string",
        "enum": [
          "active",
          "suspended",
          "closed"
        ]
      },
...
      "ModuleKey": {
        "type": "string",
        "enum": [
          "production",
          "attendance",
          "advances",
          "payroll",
          "staff",
          "machines",
          "shifts",
          "bonus_rules",
          "recycle_bin",
          "users",
          "settings"
        ],
        "description": "From the SQL catalog."
      },
...
      "ErrorCode": {
        "type": "string",
        "enum": [
          "validation_failed",
          "unauthenticated",
          "token_expired",
          "forbidden",
          "super_admin_only",
          "owner_only",
          "password_change_required",
          "property_suspended",
          "not_found",
          "machine_number_taken",
          "machine_limit_reached",
          "staff_code_taken",
          "period_already_paid",
          "restore_blocked",
          "shift_cycle_clash",
          "slab_ladder_invalid",
          "grant_exceeds_granter",
          "file_too_large",
          "rate_limited"
        ]
      },
      "Property": {
        "type": "object",
        "description": "One customer's unit — the tenant.",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64",
            "description": "Server-issued id."
          },
          "code": {
            "type": "string",
            "description": "Short handle, e.g. \"janki-creation\"."
          },
          "name": {
            "type": "string"
          },
          "timezone": {
            "type": "string",
            "description": "IANA zone; \"today\" is this zone's date."
          },
          "status": {
            "$ref": "#/components/schemas/PropertyStatus"
          },
          "machineLimit": {
            "type": "integer",
            "description": "Most machines the unit may have. Set from the web admin panel only."
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "code",
          "name",
          "timezone",
          "status",
          "machineLimit",
          "createdAt"
        ]
      },
      "PropertySettings": {
        "type": "object",
        "properties": {
          "businessName": {
            "type": "string"
          },
          "currencySymbol": {
            "type": "string"
          },
          "shiftScheme": {
            "$ref": "#/components/schemas/ShiftScheme"
          },
          "dayStartMinute": {
            "type": "integer",
            "minimum": 0,
            "maximum": 1439
          },
          "noLeaveBonusAmount": {
            "type": "number",
            "multipleOf": 0.01,
            "description": "Rupees, two decimals."
          },
          "stitchBasedEnabled": {
            "type": "boolean"
          },
          "stitchBasedRatePerThousand": {
            "type": "number",
            "description": "Up to four decimals."
          },
          "timezone": {
            "type": "string"
          }
        },
        "required": [
          "businessName",
          "currencySymbol",
          "shiftScheme",
          "dayStartMinute",
          "noLeaveBonusAmount",
          "stitchBasedEnabled",
          "stitchBasedRatePerThousand",
          "timezone"
        ]
      },
      "User": {
        "type": "object",
        "description": "A person who can sign in. Never carries a password hash.",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64",
            "description": "Server-issued id."
          },
          "displayName": {
            "type": "string"
          },
          "username": {
            "type": [
              "string",
              "null"
            ]
          },
          "phone": {
            "type": [
              "string",
              "null"
            ],
            "description": "E.164, e.g. \"+919812345678\"."
          },
          "email": {
            "type": [
              "string",
              "null"
            ]
          },
          "status": {
            "$ref": "#/components/schemas/AccountStatus"
          },
          "mustChangePassword": {
            "type": "boolean"
          },
          "lastLoginAt": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "displayName",
          "username",
          "phone",
          "email",
          "status",
          "mustChangePassword",
          "lastLoginAt"
        ]
      },
      "Membership": {
        "type": "object",
        "properties": {
          "memberId": {
            "type": "integer",
            "format": "int64",
            "description": "Server-issued id."
          },
          "propertyId": {
            "type": "integer",
            "format": "int64",
            "description": "Server-issued id."
          },
          "propertyCode": {
            "type": "string"
          },
          "propertyName": {
            "type": "string"
          },
          "role": {
            "$ref": "#/components/schemas/UserRole"
          },
          "isOwner": {
            "type": "boolean"
          },
          "status": {
            "$ref": "#/components/schemas/AccountStatus"
          }
        },
        "required": [
          "memberId",
          "propertyId",
          "propertyCode",
          "propertyName",
          "role",
          "isOwner",
          "status"
        ]
      },
      "ModuleAccess": {
        "type": "object",
        "properties": {
          "module": {
            "$ref": "#/components/schemas/ModuleKey"
          },
          "level": {
            "$ref": "#/components/schemas/AccessLevel"
          }
        },
        "required": [
          "module",
          "level"
        ]
      },
      "Member": {
        "type": "object",
        "description": "A user's place in a property.",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64",
            "description": "Server-issued id."
          },
          "userId": {
            "type": "integer",
            "format": "int64",
            "description": "Server-issued id."
          },
          "displayName": {
            "type": "string"
          },
          "role": {
            "$ref": "#/components/schemas/UserRole"
          },
          "isOwner": {
            "type": "boolean",
            "description": "The property's main super admin."
          },
          "status": {
            "$ref": "#/components/schemas/AccountStatus"
          },
          "staffId": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64",
            "description": "The roster row this login is, if any."
          },
          "modules": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ModuleAccess"
            }
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "userId",
          "displayName",
          "role",
          "isOwner",
          "status",
          "staffId",
          "modules",
          "createdAt"
        ]
      },
      "Me": {
        "type": "object",
        "properties": {
          "user": {
            "$ref": "#/components/schemas/User"
          },
          "member": {
            "$ref": "#/components/schemas/Member"
          },
          "property": {
            "$ref": "#/components/schemas/Property"
          },
          "settings": {
            "$ref": "#/components/schemas/PropertySettings"
          },
          "modules": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ModuleAccess"
            },
            "description": "Gate every screen on this list."
          }
        },
        "required": [
          "user",
          "member",
          "property",
          "settings",
          "modules"
        ]
      },
...
      "LoginInput": {
        "type": "object",
        "properties": {
          "identifier": {
            "type": "string",
            "description": "Phone (E.164), email or username."
          },
          "password": {
            "type": "string"
          },
          "client": {
            "$ref": "#/components/schemas/ClientKind"
          },
          "keepSignedIn": {
            "type": "boolean",
            "description": "15 days instead of 7."
          }
        },
        "required": [
          "identifier",
          "password",
          "client",
          "keepSignedIn"
        ]
      },
      "LoginResult": {
        "type": "object",
        "properties": {
          "accessToken": {
            "type": "string",
            "description": "JWT, 15 minutes."
          },
          "refreshToken": {
            "type": "string",
            "description": "Opaque. Rotates on every use. Keep it in the platform keystore."
          },
          "expiresIn": {
            "type": "integer",
            "description": "Seconds."
          },
          "user": {
            "$ref": "#/components/schemas/User"
          },
          "memberships": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Membership"
            }
          },
          "activeMember": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Membership"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": [
          "accessToken",
          "refreshToken",
          "expiresIn",
          "user",
          "memberships",
          "activeMember"
        ]
      },
      "TokenPair": {
        "type": "object",
        "properties": {
          "accessToken": {
            "type": "string"
          },
          "refreshToken": {
            "type": "string"
          },
          "expiresIn": {
            "type": "integer"
          }
        },
        "required": [
          "accessToken",
          "refreshToken",
          "expiresIn"
        ]
      },
      "SelectPropertyInput": {
        "type": "object",
        "properties": {
          "memberId": {
            "type": "integer",
            "format": "int64",
            "description": "Server-issued id."
          }
        },
        "required": [
          "memberId"
        ]
      },
      "RefreshInput": {
        "type": "object",
        "properties": {
          "refreshToken": {
            "type": "string"
          }
        },
        "required": [
          "refreshToken"
        ]
      },
      "ChangePasswordInput": {
        "type": "object",
        "properties": {
          "current": {
            "type": "string"
          },
          "next": {
            "type": "string",
            "description": "At least 8 characters, not the demo password."
          }
        },
        "required": [
          "current",
          "next"
        ]
      },
      "SettingsPatch": {
        "type": "object",
        "description": "The shift arrangement and day start go through /shifts/scheme.",
        "properties": {
          "businessName": {
            "type": "string"
          },
          "currencySymbol": {
            "type": "string"
          },
          "noLeaveBonusAmount": {
            "type": "number",
            "multipleOf": 0.01,
            "description": "Rupees, two decimals."
          },
          "stitchBasedEnabled": {
            "type": "boolean"
          },
          "stitchBasedRatePerThousand": {
            "type": "number",
            "description": "Up to four decimals."
          }
        }
      },
...
      "ApiError": {
        "type": "object",
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ApiErrorBody"
          }
        },
        "required": [
          "error"
        ]
      },
      "ApiErrorBody": {
        "type": "object",
        "properties": {
          "code": {
            "$ref": "#/components/schemas/ErrorCode"
          },
          "message": {
            "type": "string",
            "description": "English, for logs. Clients word `code` in the user's language."
          },
          "details": {
            "type": "object",
            "additionalProperties": {}
          }
        },
        "required": [
          "code",
          "message"
        ]
      },
      "PlatformLoginInput": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string"
          },
          "password": {
            "type": "string"
          }
        },
        "required": [
          "email",
          "password"
        ]
      },
...
    }
  },
  "paths": {
    "/v1/auth/login": {
      "post": {
        "operationId": "postAuthLogin",
        "summary": "Sign in with phone, email or username.",
        "tags": [
          "auth"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginResult"
                }
              }
            }
          },
          "default": {
            "description": "An error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        },
        "security": []
      }
    },
    "/v1/auth/select-property": {
      "post": {
        "operationId": "postAuthSelectProperty",
        "summary": "Act in one of the user's properties; also the switcher.",
        "tags": [
          "auth"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SelectPropertyInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenPair"
                }
              }
            }
          },
          "default": {
            "description": "An error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/v1/auth/refresh": {
      "post": {
        "operationId": "postAuthRefresh",
        "summary": "Rotate the refresh token.",
        "tags": [
          "auth"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenPair"
                }
              }
            }
          },
          "default": {
            "description": "An error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        },
        "security": []
      }
    },
    "/v1/auth/logout": {
      "post": {
        "operationId": "postAuthLogout",
        "summary": "Revoke this session (?everywhere=true for all).",
        "tags": [
          "auth"
        ],
        "parameters": [
          {
            "name": "everywhere",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshInput"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "Done."
          },
          "default": {
            "description": "An error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/v1/auth/change-password": {
      "post": {
        "operationId": "postAuthChangePassword",
        "summary": "Change one's own password; clears mustChangePassword.",
        "tags": [
          "auth"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangePasswordInput"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "Done."
          },
          "default": {
            "description": "An error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/v1/me": {
      "get": {
        "operationId": "getMe",
        "summary": "Who is signed in, where, and with which modules.",
        "tags": [
          "me"
        ],
        "responses": {
          "200": {
            "description": "OK.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Me"
                }
              }
            }
          },
          "default": {
            "description": "An error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/v1/settings": {
      "get": {
        "operationId": "getSettings",
        "summary": "The unit's settings.",
        "tags": [
          "settings"
        ],
        "responses": {
          "200": {
            "description": "OK.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PropertySettings"
                }
              }
            }
          },
          "default": {
            "description": "An error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "x-action": "settings.view",
        "x-destructive": false,
        "x-owner-only": false
      },
      "patch": {
        "operationId": "patchSettings",
        "summary": "Change unit settings.",
        "tags": [
          "settings"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SettingsPatch"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PropertySettings"
                }
              }
            }
          },
          "default": {
            "description": "An error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "x-action": "settings.update",
        "x-destructive": false,
        "x-owner-only": false
      }
    },
...
    "/v1/platform/auth/login": {
      "post": {
        "operationId": "postPlatformAuthLogin",
        "summary": "Platform staff sign-in.",
        "tags": [
          "platform/auth"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PlatformLoginInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenPair"
                }
              }
            }
          },
          "default": {
            "description": "An error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        },
        "security": []
      }
    },
...
  }
}